refactor: rename src to utils
