+++ src/commands/plugin_commands.ts
@@ -5,8 +5,8 @@
 
 import i18next from "i18next";
 import {RepoFrontmatter, Repository} from "../settings/interface";
-import {checkRepositoryValidity, isShared} from "../src/data_validation_test";
-import {createLink, getRepoFrontmatter} from "../src/utils";
+import {checkRepositoryValidity, isShared} from "../utils/data_validation_test";
+import {createLink, getRepoFrontmatter} from "../utils/utils";
 import GithubPublisher from "../main";
 import {
 	purgeNotesRemote,
