+++ src/main.ts
@@ -7,10 +7,10 @@ import {
 	Repository,
 } from "./settings/interface";
 import { OldSettings } from "./settings/migrate";
-import {noticeLog, verifyRateLimitAPI} from "./src/utils";
+import {noticeLog, verifyRateLimitAPI} from "./utils/utils";
 import {GithubBranch} from "./publish/branch";
 import {Octokit} from "@octokit/core";
-import {getRepoSharedKey, isShared} from "./src/data_validation_test";
+import {getRepoSharedKey, isShared} from "./utils/data_validation_test";
 import {
 	shareOneNote,
 } from "./commands/commands";
