+++ src/publish/delete.ts
@@ -9,8 +9,8 @@ import {
 } from "../settings/interface";
 import { FilesManagement } from "./files";
 import { Base64 } from "js-base64";
-import {noticeLog, trimObject, verifyRateLimitAPI } from "../src/utils";
-import {isAttachment} from "../src/data_validation_test";
+import {noticeLog, trimObject, verifyRateLimitAPI } from "../utils/utils";
+import {isAttachment} from "../utils/data_validation_test";
 import i18next from "i18next";
 
 /**
