+++ src/publish/files.ts
@@ -14,8 +14,8 @@ import {
 } from "../conversion/file_path";
 import Publisher from "./upload";
 import GithubPublisher from "../main";
-import { getRepoFrontmatter, noticeLog } from "../src/utils";
-import { isAttachment } from "../src/data_validation_test";
+import { getRepoFrontmatter, noticeLog } from "../utils/utils";
+import { isAttachment } from "../utils/data_validation_test";
 import { getAPI, Link } from "obsidian-dataview";
 
 /**
