+++ src/utils/status_bar.ts
